+++ CHANGELOG.md
@@ -36,6 +36,7 @@
 - (Feature) (Scheduler) Create Integration Profile
 - (Feature) (Scheduler) Additional types
 - (Feature) Alternative Upgrade Order Feature
+- (Feature) (Scheduler) SchedV1 Integration
 
 ## [1.2.42](https://github.com/arangodb/kube-arangodb/tree/1.2.42) (2024-07-23)
 - (Maintenance) Go 1.22.4 & Kubernetes 1.29.6 libraries